Add play/pause button

+++ tile.hpp
@@ -69,7 +69,7 @@ class Creature : public Thing{
         uint32_t size;
         // uint32_t color = 0x3C4CE7;
         uint32_t color = 0xa879fd;
-        uint16_t energy = 400;
+        float energy = 400;
 
 };
